test: move test helper fn to mocks folder
