+++ src/lib/shared/components/ui/FormElements/FormRenderer/__tests__/PasswordField.spec.tsx
@@ -3,7 +3,7 @@ import { z } from 'zod';
 import { FormRenderer } from '../FormRenderer';
 import { PasswordInput } from '../types';
 import { renderWithTheme } from '../../../../fixtures';
-import { getMockInputConfig } from './getMockInputConfig';
+import { getMockInputConfig } from '../__mocks__/getMockInputConfig';
 
 describe('Password field', () => {
     const user = userEvent.setup();
