+++ src/lib/shared/components/ui/FormElements/FormRenderer/__tests__/TelephoneField.spec.tsx
@@ -3,7 +3,7 @@ import { z } from 'zod';
 import { FormRenderer } from '../FormRenderer';
 import { TelephoneInput } from '../types';
 import { renderWithTheme } from '../../../../fixtures';
-import { getMockInputConfig } from './getMockInputConfig';
+import { getMockInputConfig } from '../__mocks__/getMockInputConfig';
 
 describe('Telephone field', () => {
     const user = userEvent.setup();
