+++ src/lib/shared/components/ui/FormElements/FormRenderer/__tests__/ToggleField.spec.tsx
@@ -5,7 +5,7 @@ import { z } from 'zod';
 import { FormRenderer } from '../FormRenderer';
 import { ToggleInput } from '../types';
 import { renderWithTheme } from '../../../../fixtures';
-import { getMockInputConfig } from './getMockInputConfig';
+import { getMockInputConfig } from '../__mocks__/getMockInputConfig';
 
 describe('Toggle Field', () => {
     const user = userEvent.setup();
